// assets/src/generator/types.ts
export type PaymentParams = {
  Iznos: string;
  ImePlatitelja: string;
  AdresaPlatitelja: string;
  SjedistePlatitelja: string;
  Primatelj: string;
  AdresaPrimatelja: string;
  SjedistePrimatelja: string;
  IBAN: string;
  ModelPlacanja: string;
  PozivNaBroj: string;
  SifraNamjene: string;
  OpisPlacanja: string;
};